Break down massive file into separate modules
- Use webpack for building
- Split file into separate modules
- Remove dependency on global variables for the modules
