+++ index.html
@@ -14,7 +14,7 @@
 	<link rel="stylesheet" href="./dist/css/fa/css/font-awesome.min.css">
 	<link rel="shortcut icon" href="./assets/img/favicon.ico" />
 </head>
-<body onload="importdata()">
+<body onload="importData()">
 	<div class="header disable-select">
 		<img src="./assets/img/logo.png">
 		<h3 class="glow">MOVELIST</h3>
