+++ other/gulpfile.js
@@ -3,10 +3,10 @@ let browserSync = require('browser-sync').create();
 let sass        = require('gulp-sass');
 let minify      = require('gulp-minify');
 let cleanCSS    = require('gulp-clean-css');
+let webpack     = require('webpack-stream');
 
 // Static Server + watching scss/html files
 gulp.task('serve', ['minify-css', 'minify-js'], function() {
-
     browserSync.init({
         server: "./../"
     });
@@ -23,19 +23,11 @@ gulp.task('sass', function() {
         .pipe(gulp.dest("./../src/css"))
         .pipe(browserSync.stream());
 });
- 
+
 gulp.task('minify-js', function() {
-  gulp.src('./../src/js/**/*.js')
-    .pipe(minify({
-        ext:{
-            src:'.js',
-            min:'.min.js'
-        },
-        ignoreFiles: ['*.min.js'],
-        preserveComments: 'some',
-        noSource: true
-    }))
-    .pipe(gulp.dest('./../dist/js/'))
+  return gulp.src('./../src/js/lib/scripts.js')
+        .pipe(webpack(require('./../webpack.config.js')))
+        .pipe(gulp.dest('./../dist/js/'));
 });
 
 gulp.task('minify-css',  ['sass'], () => {
