+++ package.json
@@ -29,12 +29,16 @@
   "devDependencies": {
     "browser-sync": "^2.18.13",
     "connect": "^3.6.2",
+    "d3": "^4.11.0",
     "gulp": "^3.9.1",
     "gulp-clean-css": "^3.7.0",
     "gulp-minify": "^1.0.0",
     "gulp-sass": "^3.1.0",
     "sass": "^1.0.0-beta.2",
     "serve-static": "^1.12.3",
+    "uglifyjs-webpack-plugin": "^1.0.0-beta.3",
+    "webpack": "^3.7.1",
+    "webpack-stream": "^4.0.0",
     "xml2js": "^0.4.17"
   }
 }
